Rename build_compilers to build_web_compilers
Closes #600

The new approach for picking up DDC (and eventually Dart2JS) will be to
have an explicit dev dependency on this package, so we want a name that
is obvious it relates to DDC (and eventuall Dart2JS).
