+++ .travis.yml
@@ -17,9 +17,9 @@ env:
   - PKG=build_barback TASK=dartanalyzer
   - PKG=build_barback TASK=dartfmt
   - PKG=build_barback TASK=test
-  - PKG=build_compilers TASK=dartanalyzer
-  - PKG=build_compilers TASK=dartfmt
-  - PKG=build_compilers TASK=test_1
+  - PKG=build_web_compilers TASK=dartanalyzer
+  - PKG=build_web_compilers TASK=dartfmt
+  - PKG=build_web_compilers TASK=test_1
   - PKG=build_config TASK=dartanalyzer
   - PKG=build_config TASK=dartfmt
   - PKG=build_config TASK=test
@@ -42,7 +42,7 @@ matrix:
     - dart: stable
       env: PKG=build_barback TASK=dartfmt
     - dart: stable
-      env: PKG=build_compilers TASK=dartfmt
+      env: PKG=build_web_compilers TASK=dartfmt
     - dart: stable
       env: PKG=build_config TASK=dartfmt
     - dart: stable
