+++ build_web_compilers/build.yaml
@@ -1,7 +1,7 @@
 builders:
   ddc:
-    target: "build_compilers"
-    import: "package:build_compilers/builders.dart"
+    target: "build_web_compilers"
+    import: "package:build_web_compilers/builders.dart"
     builder_factories:
       - moduleBuilder
       - unlinkedSummaryBuilder
@@ -19,8 +19,8 @@ builders:
     auto_apply: all_packages
     required_inputs: [".dart"]
   ddc_bootstrap:
-    target: "build_compilers"
-    import: "package:build_compilers/builders.dart"
+    target: "build_web_compilers"
+    import: "package:build_web_compilers/builders.dart"
     builder_factories: ["devCompilerBootstrapBuilder"]
     build_extensions:
       .dart:
