+++ build_web_compilers/lib/src/common.dart
@@ -7,7 +7,7 @@ import 'package:build/build.dart';
 import 'package:scratch_space/scratch_space.dart';
 
 final defaultAnalysisOptionsId =
-    new AssetId('build_compilers', 'lib/src/analysis_options.default.yaml');
+    new AssetId('build_web_compilers', 'lib/src/analysis_options.default.yaml');
 
 String defaultAnalysisOptionsArg(ScratchSpace scratchSpace) =>
     '--options=${scratchSpace.fileFor(defaultAnalysisOptionsId).path}';
