+++ build_web_compilers/lib/src/module_builder.dart
