+++ build_web_compilers/lib/src/modules.dart
