+++ build_web_compilers/lib/src/scratch_space.dart
